% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{crosscheck}
\alias{crosscheck}
\title{crosscheck.GameDayPlays}
\usage{
crosscheck(data)
}
\arguments{
\item{data}{An MLBAM data set}
}
\value{
The ratio of the Frobenius norm of the matrix of differences to the Frobenius norm of the matrix
defined by the Lahman database.
}
\description{
Cross-check the accuracy of the GameDay data with the Lahman database
}
\details{
Cross-checks summary statistics with the Lahman database.
}
\examples{
ds = getData()
crosscheck(ds)
}

