% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldingplot.R
\name{fieldingplot}
\alias{fieldingplot}
\alias{fieldingplot.default}
\alias{fieldingplot.formula}
\alias{fieldingplot.glm}
\alias{fieldingplot.lm}
\alias{panel.fielding}
\title{fieldingplot}
\usage{
fieldingplot(x, data, ...)

\method{fieldingplot}{glm}(x, data, ...)

\method{fieldingplot}{lm}(x, data, ...)

\method{fieldingplot}{formula}(x, data, label = "label", write.pdf = FALSE,
  ...)

\method{fieldingplot}{default}(x, y, z, label = "label", write.pdf = FALSE,
  ...)

panel.fielding(x, y, z, ...)
}
\arguments{
\item{x}{a model object, currently \code{\link{lm}} and \code{\link{glm}} are
supported. For the default method this is an x-coordinate to be passed to
\code{\link{contourplot}}.}

\item{data}{a \code{\link{GameDayPlays}} dataset}

\item{...}{arguments passed to \code{\link{panel.contourplot}}}

\item{label}{a string to be used in the resulting filename}

\item{write.pdf}{a \code{logical} indicating whether to write the contourplot
to a file}

\item{y}{y-coordinate for default method passed to \code{\link{contourplot}}.}

\item{z}{z-coordinate for default method passed to \code{\link{contourplot}}.}
}
\value{
A contourplot object
}
\description{
Creates a plot of the fielding model for an individual player
}
\details{
This is a convenience function for visually checking the fielding models.
It provides a way to inspect the quality of the fielding models for individual
players, and was used to generate one of the figures used in the \code{\link{openWAR}}
paper. However, this function is not exported, since it is not really accessible
to end users. We hope to add this functionality in a future release.
}
\examples{
deltas <- makeWARre24(May)
Mayplus = cbind(May, deltas)
library(dplyr)
BIP <- filter(Mayplus, isBIP == TRUE)
\dontrun{
fielding = makeWARFielding(BIP)
}
}

