% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlaysExt.R
\docType{class}
\name{GameDayPlaysExt-class}
\alias{GameDayPlaysExt-class}
\title{GameDayPlaysExt}
\description{
Contains the output from \code{\link{makeWAR}}
}
\examples{
showClass('GameDayPlaysExt')
}

