% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getDataWeekly}
\alias{getDataWeekly}
\title{getDataWeekly}
\usage{
getDataWeekly(start = Sys.Date() - 8)
}
\arguments{
\item{start}{A valid date in yyyy-mm-dd format (default Sys.Date()-8)}
}
\value{
A data.frame of class "GameDayPlays" consisting of play-by-play data
}
\description{
Retrieves MLBAM GameDay files for a single week
}
\details{
Given a date, this function will retrieve data from the week starting on the specified date from the
GameDay server and process them into a single data.frame.
}
\examples{
#Retrieve all the data from the first week of the 2013 season
ds = getDataWeekly("2013-03-31")
}

