% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeWAR.R
\name{getFielderResp}
\alias{getFielderResp}
\title{getFielderResp}
\usage{
getFielderResp(data, ...)
}
\arguments{
\item{data}{An MLBAM data.frame}

\item{...}{currently ignored}
}
\value{
data.frame with 9 columns, each row representing a ball in play
}
\description{
Find the shared responsibility for balls in play
}
\details{
Fits 9 logistic regression models, each giving the probability of
a fielder at one of the 9 defensive positions successfully converting the
ball into at least one out.
}

