\name{getGameDay}
\alias{getGameDay}
\title{getGameDay}
\usage{
  getGameDay(gameId = "gid_2012_08_12_atlmlb_nynmlb_1")
}
\arguments{
  \item{gameId}{A valid MLBAM gameId}
}
\value{
  Nothing, but writes files to disk
}
\description{
  Retrieves MLBAM GameDay files from the Internet
}
\details{
  A variety of functions for downloading XML files are
  provided
}

