% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModels.R
\name{getMatrixRunExpectancy}
\alias{getMatrixRunExpectancy}
\title{getMatrixRunExpectancy}
\usage{
getMatrixRunExpectancy(x, ...)
}
\arguments{
\item{x}{a run expectancy model}

\item{...}{currently ignored}
}
\value{
a matrix in the familiar 8x3 form
}
\description{
a untility function for retrieving a run expectancy
model in matrix form.
}
\examples{

ERM <- getMatrixRunExpectancy(getModelRunExpectancy(May, verbose = FALSE))
dim(ERM)
ERM
}
