% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getModels.R
\name{getModelFieldingCollective}
\alias{getModelFieldingCollective}
\title{getModelFieldingCollective}
\usage{
getModelFieldingCollective(data)
}
\arguments{
\item{data}{An MLBAM data.frame}
}
\value{
a vector representing the probability that each ball in play will be fielded
}
\description{
Determine the responsibility of the fielders, collectively
}
\details{


Computes a 2D kernel smoothed estimate of the probability that *any* of the 9 fielders
will make a play on a ball in play
}

