% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRunEx.R
\name{getRunEx}
\alias{getRunEx}
\title{getRunEx}
\usage{
getRunEx(data, drop.incomplete = TRUE, ...)
}
\arguments{
\item{data}{an MLBAM dataset}

\item{drop.incomplete}{a LOGICAL indicating whether incomplete innings (e.g. walk-off innings)
should be excluded from the run expectancy model.  Default is TRUE.}

\item{...}{currently ignored}
}
\value{
A function that takes two arguments: baseCode and outs, and returns a run value
}
\description{
Find the Run Expectancy for a given state
}
\details{
Given a data set, returns a function giving the run expectancy for any base/out state
}
\examples{
#Data from May 2013
data(May)
fit.rem = getRunEx(May)

fit.rem(0,0)
fit.rem(4,1)
}

