% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDay.R
\name{getRunnerMovement}
\alias{getRunnerMovement}
\title{getRunnerMovement}
\usage{
getRunnerMovement(x)
}
\arguments{
\item{x}{a character vector of runner movements}
}
\description{
getRunnerMovement
}

