% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDay.R
\name{getURLs.gameday}
\alias{getURLs.gameday}
\alias{getXML.gameday}
\title{getURLs.gameday}
\usage{
getURLs.gameday(gd)
}
\arguments{
\item{gd}{A \code{gameday} object}
}
\value{
a vector of URLs
}
\description{
Convenience function for returning data related to a game
}
\details{
Given a gameId string, this function can return several URLs
}

