% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDay.R
\name{getURLs}
\alias{getURLs}
\title{getURLs.gameday}
\usage{
getURLs(gd, ...)
}
\arguments{
\item{gd}{A \code{gameday} object}

\item{...}{currently ignored}
}
\value{
a vector of URLs
}
\description{
Convenience function for returning data related to a game
}
\details{
Given a gameId string, this function can return several URLs
}
