% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeWAR.R
\name{makeWARBaserunning}
\alias{makeWARBaserunning}
\title{makeWARBaserunning}
\usage{
makeWARBaserunning(data, ...)
}
\arguments{
\item{data}{An MLBAM data.frame}

\item{...}{currently ignored}
}
\value{
A data.frame with 4 columns: baserunner id, raa for the runner on first base, raa for the runner on second base, and raa for the runner on third base.
}
\description{
Find the shared responsibility models for balls in play
}
\details{
blah
}

