% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWAR.R
\name{makeWARFielding}
\alias{makeWARFielding}
\title{Compute RAA for fielders.}
\usage{
makeWARFielding(data, ...)
}
\arguments{
\item{data}{A \code{GameDayPlays} object}

\item{...}{currently ignored}
}
\value{
A data.frame with 12 columns: \code{p.hat} contains the probability that any fielder makes an out given the location 
of the ball in play; \code{delta.field} is the total delta assigned to all of the fielders; columns 3 through 11 
contains the delta values for each od the 9 fielders (i.e. these columns sum to \code{delta.field}); \code{venueId} contains information about
the stadium that the plate appearance took place in.
}
\description{
Compute the RAA values for all of the fielders.  Used internally in \code{makeWAR}.
}
