% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWAR.R
\name{makeWARre24}
\alias{makeWARre24}
\title{Compute RE24 before and after each PA.}
\usage{
makeWARre24(data, mod.re = NULL, ...)
}
\arguments{
\item{data}{A \code{GameDayPlays} object}

\item{mod.re}{an \code{lm} object giving the Run Expectancy Model. If 
NULL (the default), the run expectancy model will be generated from \code{data}}

\item{...}{currently ignored}
}
\value{
a data frame with three columns: \code{startExR}, \code{endExR}, and 
\code{delta}. The latter is the change in run expectancy.
}
\description{
Compute the Run Expectancy at the beginning and end of every play
in a data frame, as well as their difference (RE24).
}
\examples{

Maymod.re = getModelRunExpectancy(May)
re24 <- makeWARre24(May14)
re24full <- makeWARre24(May14, mod.re = Maymod.re)
sum(re24$delta)
sum(re24full$delta)
}
