% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{May14}
\alias{May14}
\title{A sample of GameDayPlays data from MLBAM}
\format{A \code{\link{GameDayPlays}} object, which is also a \code{data.frame} with 1132
rows.}
\source{
\url{http://gd2.mlb.com/components/game/mlb/year_2012/month_05/day_14/}
}
\usage{
May14
}
\description{
A dataset containing play-by-play data from MLBAM GameDay and processed by
\code{openWAR} for all games on May 14, 2013.
}
\keyword{datasets}

