% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MayProcessed}
\alias{MayProcessed}
\title{A sample of processed data that has been through \code{\link{makeWAR}}}
\format{A list with four components:
\describe{
   \item{plays}{the original \code{\link{GameDayPlays}} data on which the models were fit. NULL in this case}
   \item{data}{NULL}
   \item{models.used}{a list of model objects used in the computation of \code{\link{makeWAR}}}
   \item{openWAR}{an \code{\link{openWARPlays}} (\code{\link{data.frame}}) with the resulting processed values, based on the openWAR model}
   }}
\usage{
MayProcessed
}
\description{
A sample of processed data that has been through \code{\link{makeWAR}}
}
\keyword{datasets}
