% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetsBraves}
\alias{MetsBraves}
\title{A sample of play-by-play data from MLBAM}
\format{See \code{\link{GameDayPlays}}}
\source{
\url{http://gd2.mlb.com/components/game/mlb/year_2012/month_08/day_12/gid_2012_08_12_atlmlb_nynmlb_1/}
}
\usage{
MetsBraves
}
\description{
A dataset containing play-by-play data from MLBAM GameDay and processed by
\code{openWAR} for the August 8th, 2012 game between the New York Mets and Atlanta Braves.
}
\keyword{datasets}

