% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{openWAR}
\alias{openWAR}
\alias{openWAR-package}
\title{openWAR}
\description{
This package allows a user to explore WAR
}
\details{
A fully open-source implementation of WAR.
}
\author{
Ben Baumer and Gregory J. Matthews
}

