% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlays.R
\docType{class}
\name{openWARPlays-class}
\alias{openWARPlays-class}
\title{openWARPlays}
\description{
Contains the output from the function \code{\link{makeWAR}}
specific to \code{openWAR}.
}
\examples{
showClass('openWARPlays')
}

