% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlays.R
\name{openWARPlays}
\alias{openWARPlays}
\title{openWARPlays}
\usage{
openWARPlays(x)
}
\arguments{
\item{x}{an object}
}
\description{
Contains the output from the function \code{\link{makeWAR}}
specific to \code{openWAR}.
}
\examples{
class(MayProcessed$openWAR)
}
