% Generated by roxygen2 (4.0.1): do not edit by hand
\name{panel.baseball}
\alias{panel.baseball}
\title{panel.baseball}
\usage{
panel.baseball()
}
\value{
nothing
}
\description{
Visualize Balls in Play
}
\details{
A convenience function for drawing a generic baseball field using a Cartesian coordinate
system scaled in feet with home plate at the origin.
}
\examples{
ds = getData()
plot(ds)
}

