% Generated by roxygen2 (4.0.1): do not edit by hand
\name{panel.fielding}
\alias{panel.fielding}
\title{fieldingplot}
\usage{
panel.fielding(x, y, z, ...)
}
\arguments{
\item{data}{a GameDayPlays dataset}

\item{model}{a model object that has a `predict` method}

\item{write.pdf}{a LOGICAL indicating whether to write the contourplot to a file}
}
\value{
A contourplot object
}
\description{
Plot the fielding model for an individual player
}
\details{
This is a convenience function for visually checking the fielding models.
}
\examples{
data(MLBAM2013)
# not run
}

