% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{panel.baseball}
\alias{panel.baseball}
\alias{plot.GameDayPlays}
\title{panel.baseball}
\usage{
panel.baseball()

\method{plot}{GameDayPlays}(x, batterName = NULL, pitcherName = NULL,
  event = NULL, pch = 1, ...)
}
\arguments{
\item{x}{A GameDayPlays data set with fields 'our.x' and 'our.y'}

\item{batterName}{A character string containing the last name of a batter}

\item{pitcherName}{A character string containing the last name of a pitcher}

\item{pch}{A numeric}
}
\value{
nothing

an xyplot()
}
\description{
Visualize Balls in Play

Visualize Balls in Play
}
\details{
A convenience function for drawing a generic baseball field using a Cartesian coordinate
system scaled in feet with home plate at the origin.

Plots the balls in play from GameDay data. This function will plot (x,y)-coordinates
with a generic baseball field plotted in the background. Other lattice options can be passed
to xyplot().
}
\examples{
plot(May)
}

