% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{plot.openWARPlayers}
\alias{plot.openWARPlayers}
\title{plot.openWARPlayers}
\usage{
\method{plot}{openWARPlayers}(x, ...)
}
\arguments{
\item{x}{A data.frame object of class \code{'openWARPlayers'}}
}
\description{
Display a season's worth of openWAR results
}
\details{
Given an openWARPlayers object, draw a plot displaying each player's RAA, WAR, and replacement
level shadow.
}
\examples{
out = makeWAR(May)
players = getWAR(out$openWAR)
summary(players)
plot(players)
}

