% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDay.R
\name{readData.gameday}
\alias{readData.gameday}
\title{get game data}
\usage{
readData.gameday(gd)
}
\arguments{
\item{gd}{A gameday object}
}
\value{
A gameday object
}
\description{
Grabs XML files from MLBAM and converts to data.frames in memory
}
\details{
Grabs XML files from MLBAM and converts to data.frames in memory
}
\examples{
library(openWAR)
gd <- gameday()
}

