% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDay.R
\name{readData}
\alias{readData}
\title{get game data}
\usage{
readData(gd, ...)
}
\arguments{
\item{gd}{A gameday object}

\item{...}{currently ignored}
}
\value{
A gameday object
}
\description{
Grabs XML files from MLBAM and converts to data.frames in memory
}
\details{
Grabs XML files from MLBAM and converts to data.frames in memory
}
\examples{

library(openWAR)
gd <- gameday()
}
