% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{shakeWAR}
\alias{shakeWAR}
\title{shakeWAR}
\usage{
shakeWAR(data, resample = "plays", N = 10, ...)
}
\arguments{
\item{data}{An MLBAM data.frame}

\item{resample}{An element of \code{c('plays', 'models', 'both')}}

\item{N}{the number of resamples (default 5000)}
}
\value{
a data.frame with RAA values
}
\description{
resample a data.frame to obtain variance estimate for WAR
}
\details{
Resamples the rows of an MLBAM data set
}
\examples{
ds = getData()
res = shakeWAR(ds, resample='plays', N=10)
summary(res)
}

