% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{summary.GameDayPlays}
\alias{summary.GameDayPlays}
\title{summary.GameDayPlays}
\usage{
\method{summary}{GameDayPlays}(object, ...)
}
\arguments{
\item{object}{A \code{\link{GameDayPlays}} data set}

\item{...}{currently ignored}
}
\description{
Summarize MLBAM data
}
\details{
Prints information about the contents of an object of class \code{\link{GameDayPlays}}
}
\examples{
summary(May)
}

