\name{summary.MLBAMData}
\alias{summary.MLBAMData}
\title{summary.MLBAMData}
\usage{
  summary.MLBAMData(data)
}
\arguments{
  \item{data}{An MLBAMData set}
}
\value{
  nothing
}
\description{
  Summarize MLBAM data
}
\details{
  Prints information about the contents of an MLBAMData
  set.
}
\examples{
ds = getData()
summary(ds)
}

