% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{tabulate}
\alias{tabulate}
\alias{tabulate.GameDayPlays}
\title{tabulate}
\usage{
tabulate(data)

\method{tabulate}{GameDayPlays}(data)
}
\arguments{
\item{data}{An object of class \code{\link{GameDayPlays}}}
}
\value{
A data.frame of seasonal totals for each team
}
\description{
Summarize MLBAM data
}
\details{
Tabulates Lahman-style statistics by team for the contents of a \code{\link{GameDayPlays}} data set.
}
\note{
Sean Lahman database <http://www.seanlahman.com/baseball-archive/statistics/>
}
\examples{
library(openWAR)
tabulate(May)
}

