% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{updateGame}
\alias{updateGame}
\title{updateGame}
\usage{
updateGame(gameId.vec, data, ...)
}
\arguments{
\item{data}{a data.frame returned by getData()}

\item{gameId}{A valid MLBAM gameId}
}
\value{
a data.frame
}
\description{
Replaces data from a single game
}
\details{
Deletes, and then rbinds fresh information from a particular game.
}
\examples{
getData(start = "2013-03-31", end = "2013-07-14")
ds = updateGame("gid_2013_04_16_nynmlb_colmlb_2")
}

