% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDay.R
\name{updateHalfInning}
\alias{updateHalfInning}
\title{updataHalfInning}
\usage{
updateHalfInning(dat)
}
\arguments{
\item{dat}{a \code{GameDayPlays} object containing information about a single inning.}
}
\value{
A gameday object
}
\description{
updateHalfInning
}

