% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscheck.R
\name{crosscheck}
\alias{crosscheck}
\alias{crosscheck.GameDayPlays}
\title{crosscheck}
\usage{
crosscheck(data)

\method{crosscheck}{GameDayPlays}(data)
}
\arguments{
\item{data}{\code{GameDayPlays} data frame}
}
\value{
The ratio of the Frobenius norm of the matrix of differences to the Frobenius norm of the matrix
defined by the Lahman database.
}
\description{
Cross-check the accuracy of the GameDay data with the Lahman database
}
\details{
Cross-checks summary statistics with the Lahman database.
}
