% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDay.R
\name{gameday}
\alias{gameday}
\title{Convenience class for dealing with MLBAM GameDay files}
\usage{
gameday(gameId = "gid_2012_08_12_atlmlb_nynmlb_1", ...)
}
\arguments{
\item{gameId}{A valid MLBAM gameId}

\item{...}{currently ignored}
}
\value{
An object of class \code{\link{gameday}}, which consists of a list containing
\item{gameId}{The ID of the game (e.g. 'gid_2012_08_12_atlmlb_nynmlb_1')}
\item{base}{The base URL for the game (e.g. 'http://gd2.mlb.com/components/game/mlb/year_2012/month_08/day_12/')}
\item{url}{A character vector containing the names of the XML files downloaded from the GameDay server}
\item{ds}{The \code{\link{GameDayPlays}} object for use with openWAR containing all 
play-by-play information for the game indicated by the gameId}
}
\description{
Each gameId corresponds to one object of class \code{\link{gameday}}
}
\details{
A \code{\link{gameday}} object is a list containing 4 elements: \code{gameId}, \code{base}, \code{url}, and \code{ds}.  \code{gameId} is the 
unique identifier for a game, \code{base} contains the base URL corresponding the the \code{gameId}, \code{url} contains the sub urls pointing to different 
data sets that are combined together to create an object of class \code{\link{GameDayPlays}} containing play-by-play data.  The \code{ds} slot of the list 
contains the \code{\link{GameDayPlays}} object of play-by-play data.
}
\examples{

exampleGameday <- gameday(gameId = 'gid_2012_08_12_atlmlb_nynmlb_1')
# Display GameId
exampleGameday$gameId
# Summary of the different types of game events for gameId = 'gid_2012_08_12_atlmlb_nynmlb_1'
summary(exampleGameday$ds$event)
plot(exampleGameday$ds)
}
