% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(start = Sys.Date() - 1, end = NULL, gameIds = NULL,
  drop.suspended = TRUE)
}
\arguments{
\item{start}{A valid date in yyyy-mm-dd format (default today)}

\item{end}{A valid date in yyyy-mm-dd format (default start)}

\item{drop.suspended}{Logical indicating whether games with fewer than 5 innings should be excluded}
}
\value{
A data.frame of class "GameDayPlays" consisting of play-by-play data
}
\description{
Retrieves MLBAM GameDay files for a specified time interval using multiple calls of the gameday function.
}
\details{
Given a beginning and end date, this function will retrieve all data from the MLABM
GameDay server in the specified interval and process them into a single data.frame.
}
\examples{
# Get data from one day
ds = getData(start = "2013-03-31")
# Get data from multiple days
ds = getData(start = "2013-03-31", end = "2013-04-02")
# Get data for entire 2013 season
# ds = getData(start = "2013-03-31", end = "2013-09-30")
}

