% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(start = Sys.Date() - 1, end = NULL, gameIds = NULL,
  drop.suspended = TRUE)
}
\arguments{
\item{start}{A valid date in yyyy-mm-dd format (default is \code{Sys.Date() - 1} (i.e. yesterday))}

\item{end}{A valid date in yyyy-mm-dd format (default is \code{NULL}, which then uses the same value as \code{start} (i.e. yesterday))}

\item{gameIds}{a vector of specific gameIds that you want to retrieve. If
NULL (the default), then the dates are used to fetch the relevant gameIds}

\item{drop.suspended}{Logical indicating whether games with fewer than 5 innings should be excluded.  Default is \code{TRUE}.}
}
\value{
A data.frame of class 'GameDayPlays' consisting of play-by-play data
}
\description{
Retrieves MLBAM GameDay files for a specified time interval using multiple calls of the gameday function.
}
\details{
Given a beginning and end date, this function will retrieve all data from the MLABM
GameDay server in the specified interval and process them into a single data.frame.
}
\examples{
# Get data from one day
ds = getData(start = '2013-03-31')
# Get data from multiple days
\dontrun{
ds = getData(start = '2013-03-31', end = '2013-04-02')
}
# Get data for entire 2013 season
\dontrun{
ds = getData(start = '2013-03-31', end = '2013-09-30')
}
}

