% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getDataMonthly}
\alias{getDataWeekly}
\title{Retrieves MLBAM GameDay files}
\usage{
getData(start = Sys.Date() - 1, end = NULL, gameIds = NULL,
  drop.suspended = TRUE)

getDataMonthly(yyyy = 2013, m = 5)

getDataWeekly(start = Sys.Date() - 8)
}
\arguments{
\item{start}{A valid date in yyyy-mm-dd format (default is \code{\link{Sys.Date}() - 1} (i.e. yesterday))}

\item{end}{A valid date in yyyy-mm-dd format (default is \code{NULL}, which then uses the same value as \code{start} (i.e. yesterday))}

\item{gameIds}{a vector of specific gameIds that you want to retrieve. If 
NULL (the default), then the dates are used to fetch the relevant gameIds}

\item{drop.suspended}{Logical indicating whether games with fewer than 5 innings should be excluded.  Default is \code{TRUE}.}

\item{yyyy}{A year}

\item{m}{A numeric value corresponding to a month}
}
\value{
A \code{\link{tbl_df}} of class \code{\link{GameDayPlays}} consisting of play-by-play data
}
\description{
Retrieves MLBAM GameDay files for a specified time interval using
 multiple calls of the \code{\link{gameday}} function.
}
\details{
Given a beginning and end date, this function will retrieve all data from the MLBAM
GameDay server in the specified interval and process them into a single data frame.
}
\examples{

# Get data from one day
ds = getData(start = '2013-03-31')
# Get data from multiple days
\dontrun{
ds = getData(start = '2013-03-31', end = '2013-04-02')
}
# Get data for entire 2013 season
\dontrun{
ds = getData(start = '2013-03-31', end = '2013-09-30')
}

# Retrieve all of the MLBAM data from May 2013
\dontrun{
# ds = getDataMonthly(2013, 5)
}

# Retrieve all the data from the first week of the 2013 season
\dontrun{
ds = getDataWeekly('2013-03-31')
}
}
