% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getGameIds}
\alias{getGameIds}
\title{Retrives gameIds.}
\usage{
getGameIds(date = Sys.Date())
}
\arguments{
\item{date}{A date in 'yyyy-mm-dd' format}
}
\value{
A vector of gameIds
}
\description{
Retrieves MLBAM gameIds for a specified date.
}
\details{
Downloads information for a given day from the MLBAM website and retrieves
a list of valid gameIds. This function is used internally in the function \code{getData}.
}
\examples{
getGameIds()
getGameIds('2008-05-14')

}
