% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getModels.R
\name{getModels}
\alias{getModels}
\alias{getModels.GameDayPlays}
\title{getModels}
\usage{
getModels(data, ...)
}
\arguments{
\item{data}{a GameDayPlays dataset}

\item{type}{the type of model to be built.}

\item{drop.incomplete}{a LOGICAL indicating whether incomplete innings (e.g. walk-off innings)
should be excluded}
}
\value{
A model object with a predict() method
}
\description{
Retrieve various models trained on GameDayPlays data
}
\details{
This function will build various models based on the MLBAM data set and the openWAR framework.
}
\examples{
data(May)
re.mod = getModels(May, type = 'run-expectancy')

# Generate the Run Expectancy Matrix
states = expand.grid(startCode = 0:7, startOuts = 0:2)
matrix(predict(re.mod[["run-expectancy"]], newdata=states), ncol=3)
}

