% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlays.R
\name{getRAA}
\alias{getRAA}
\alias{getRAA.openWARPlays}
\title{getRAA}
\usage{
getRAA(data, ...)

\method{getRAA}{openWARPlays}(data, ...)
}
\arguments{
\item{data}{An openWARPlays data.frame}

\item{...}{currently ignored}
}
\value{
an RAA object
}
\description{
Tabulates RAA values per player
}
\details{
takes an openWARPlays data.frame and consolidates the RAA values by player
}
\examples{
players = getRAA(MayProcessed$openWAR)
}

