% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlays.R
\name{getRAA}
\alias{getRAA}
\title{getRAA}
\usage{
getRAA(data, ...)
}
\arguments{
\item{data}{An openWARPlays data.frame}
}
\value{
an RAA object
}
\description{
Tabulates RAA values per player
}
\details{
takes an openWARPlays data.frame and consolidates the RAA values by player
}
\examples{
ds = makeWAR(May)
players = getRAA(ds$openWAR)
summary(players)
}

