% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlays.R
\name{getRAA}
\alias{getRAA}
\alias{getRAA.openWARPlays}
\title{getRAA}
\usage{
getRAA(data, ...)

\method{getRAA}{openWARPlays}(data, ...)
}
\arguments{
\item{data}{An object of class \code{\link{openWARPlays}}}

\item{...}{currently ignored}
}
\value{
A data.frame of class \code{\link{openWARPlayers}}
}
\description{
Tabulates runs above average (RAA) values per player
}
\details{
Takes an object of class \code{\link{openWARPlays}} and consolidates the RAA values by player
}
