% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{getReplacementActivity}
\alias{getReplacementActivity}
\alias{getReplacementActivity.openWARPlays}
\title{getReplacementActivity}
\usage{
getReplacementActivity(data, replacementIds)

\method{getReplacementActivity}{openWARPlays}(data, replacementIds)
}
\arguments{
\item{data}{A data.frame of class \code{\link{openWARPlays}}}

\item{replacementIds}{A vector of playerIds for replacement-level players}
}
\value{
A list of RAA values for each type of activity that generates RAA values
}
\description{
Isolates the plays involving replacement-level players
}
\details{
Returns only the RAA values involving replacement-level players
}
\note{
This function is used internally in the function \code{getWAR}
}
\examples{

players = getWAR(MayProcessed$openWAR)
summary(players)
replIds = getReplacementPlayers(players)
replacementActivity = getReplacementActivity(MayProcessed$openWAR,replIds)

}
