% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replacement.R
\name{getReplacementMeans}
\alias{getReplacementMeans}
\alias{getReplacementMeans.openWARPlays}
\title{getReplacementMeans}
\usage{
getReplacementMeans(data, replIds)

\method{getReplacementMeans}{openWARPlays}(data, replIds)
}
\arguments{
\item{data}{An openWARPlays data.frame}

\item{replIds}{MLBAM IDs of specific players that you want to designate as replacement level}
}
\value{
a vector of mean contributions per activity for replacement-level players
}
\description{
Compute the mean RAA for replacement-level players
}
\details{
Identifies replacement-level players, and then compute their contribution
}
\examples{
raa <- getRAA(MayProcessed$openWAR)
replIds = getReplacementPlayers(raa, 30)
getReplacementMeans(MayProcessed$openWAR, replIds)
}

