% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replacement.R
\name{getReplacementMeans}
\alias{getReplacementMeans}
\title{getReplacementMeans}
\usage{
getReplacementMeans(data, replIds)
}
\arguments{
\item{data}{An openWARPlays data.frame}
}
\value{
a vector of mean contributions per activity for replacement-level players
}
\description{
Compute the mean RAA for replacement-level players
}
\details{
Identifies replacement-level players, and then compute their contribution
}
\examples{
ds = makeWAR(May)
raa <- getRAA(ds$openWAR)
replIds = getReplacementPlayers(raa, 30)
getReplacementMeans(ds$openWAR, replIds)
}

