% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{getReplacementMeans}
\alias{getReplacementMeans}
\alias{getReplacementMeans.openWARPlays}
\title{getReplacementMeans}
\usage{
getReplacementMeans(data, replIds)

\method{getReplacementMeans}{openWARPlays}(data, replIds)
}
\arguments{
\item{data}{A data.frame of class \code{\link{openWARPlays}}}

\item{replIds}{MLBAM IDs of specific players that you want to designate as replacement-level}
}
\value{
A numeric vector of mean contributions per activity for replacement-level players
}
\description{
Compute the mean RAA for replacement-level players
}
\details{
Identifies replacement level players and then computes their mean contribution
}
\examples{


players = getWAR(MayProcessed$openWAR)
summary(players)
replIds = getReplacementPlayers(players)
replMeans = getReplacementMeans(MayProcessed$openWAR, replIds)

}
