% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replacement.R
\name{getReplacementPlayers}
\alias{getReplacementPlayers}
\alias{getReplacementPlayers.RAA}
\title{getReplacementPlayers}
\usage{
getReplacementPlayers(data, nteams = 30, ...)
}
\arguments{
\item{data}{An object of class \code{'RAA'}}
}
\description{
This function determines and returns a list of players who are considered to be replacement level players.
}
\details{
In the openWAR framework, we first define a specific number of players to be considered 'true' major league talent.  This function returns a list of players who do not appear on the list of 'true' major league talent.
The performances of the players on this list are used to estimate the value of a replacement level player for use in converting runs above average into runs above replacement and ultimately wins above replacement.
}
\examples{
out = makeWAR(May)
players = getWAR(out$openWAR)
summary(players)
replacementIds = getReplacementPlayers(players)
}

