% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunEx.R
\name{getRunEx}
\alias{getRunEx}
\title{Computes run expectancy.}
\usage{
getRunEx(data, drop.incomplete = TRUE, ...)
}
\arguments{
\item{data}{an MLBAM dataset}

\item{drop.incomplete}{a LOGICAL indicating whether incomplete innings (e.g. walk-off innings)
should be excluded from the run expectancy model.  Default is TRUE.}

\item{...}{currently ignored}
}
\value{
A function that takes two arguments: baseCode and outs, and returns a run value
}
\description{
Find the Run Expectancy for a given state.
}
\details{
Given a data set, returns a function giving the run expectancy for any base/out state
}
\note{
baseCode takes on integer values between 0 and 7 with each value correpsonding to a unique 
base state (e.g. runner on second, runners on first and third, bases loaded).  The baseCode is the binary representation of the
 base state with first, second, and third base represented by 1, 2, and 4, respectively. 
  For example, runners on second and third corresponds to a baseCode of 6 (i.e. 2 + 4 = 6).
}
\examples{

#Data from May 2013
data(May)
fit.rem = getRunEx(May)

fit.rem(0,0)
fit.rem(4,1)

}
