% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDay.R
\name{getRunnerMovement}
\alias{getRunnerMovement}
\title{getRunnerMovement}
\usage{
getRunnerMovement(x)
}
\arguments{
\item{x}{a character vector of runner movements}
}
\description{
Condense the GameDay descriptions of runner movement into
a character vector.
}
