% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getWAR.R
\name{getWAR}
\alias{getWAR}
\alias{getWAR.GameDayPlays}
\alias{getWAR.list}
\alias{getWAR.openWARPlays}
\title{getWAR}
\usage{
getWAR(data, dataRepl = NULL, nteams = 30, verbose = TRUE, ...)

\method{getWAR}{GameDayPlays}(data, dataRepl = NULL, nteams = 30,
  verbose = TRUE, ...)

\method{getWAR}{list}(data, dataRepl = NULL, nteams = 30, verbose = TRUE,
  ...)

\method{getWAR}{openWARPlays}(data, dataRepl = NULL, nteams = 30,
  verbose = TRUE, ...)
}
\arguments{
\item{data}{An object of class \code{'openWARPlays'}}

\item{dataRepl}{An object of class \code{'openWARPlays'} that will be used to calculate the replacement level.}

\item{nteams}{The nteams argument to be passed to getReplacementPlayers}

\item{verbose}{do you want information about how many replacement-level players there are?}

\item{...}{currently ignored}
}
\description{
Tabulates WAR
}
\details{
Compute each player's WAR, given their RAA values
}
\examples{
\dontrun{
MayProcessed = makeWAR(May)
}
war = getWAR(MayProcessed)
war = getWAR(MayProcessed, nteams = 27.5)
\dontrun{
war = getWAR(May)
}
\dontrun{
MayProcessed = makeWAR(May)
}
war = getWAR(MayProcessed)
\dontrun{
MayProcessed = makeWAR(May)
}
war = getWAR(MayProcessed$openWAR)
}

