% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{May14}
\alias{May14}
\title{A sample of GameDayPlays data from MLBAM}
\format{A `GameDayPlays` object, which is also a \code{data.frame} with 1132
rows (e.g. plays) and 62 columns. The columns are as follows:
\describe{
        \item{pitcherId}{The MLBAM id of the pitcher}
       \item{batterId}{The MLBAM id of the batter}
        \item{field_teamId}{The MLBAM id of the fielding team}
        \item{ab_num}{The chronological number of the plate appearance}
        \item{inning}{The inning number}
        \item{half}{Indicates which half of the inning (i.e. top or bottom)}
        \item{balls}{The number of balls at the end of the plate appearance}
        \item{strikes}{The number of strikes at the end of the plate appearance}
        \item{endOuts}{The number of outs at the end of the plate appearance}
        \item{event}{The result of the plate appearance (e.g. Single, Walk, Grounded Into DP)}
        \item{actionID}{NA}
        \item{description}{A description of the result of the plate appearance}
        \item{stand}{The handedness of the batter during the plate appearance}
        \item{throws}{The handedness of the pitcher}
        \item{runnerMovement}{A coded description of movement of all base runners during the plate appearance}
        \item{x}{For balls in play, the MLBAM x-coordinate of where the ball is first touched. (NA for balls not in play)}
        \item{y}{For balls in play, the MLBAM y-coordinate of where the ball is first touched. (NA for balls not in play)}
        \item{game_type}{Game type}
        \item{home_team}{The MLBAM home team abbreviation}
        \item{home_teamId}{The MLBAM id of the home team}
        \item{home_lg}{The league of the home team (i.e. AL or NL)}
        \item{away_team}{The MLBAM away team abbreviation}
        \item{away_teamId}{The MLBAM id of the away team}
        \item{away_lg}{The league of the away team (i.e. AL or NL)}
        \item{venueId}{The MLBAM park id}
        \item{stadium}{The name of the stadium where the game is being played}
        \item{timestamp}{The timestamp associated withe the plate appearance}
        \item{playerId.C}{The MLBAM of the catcher}
        \item{playerId.1B}{The MLBAM of the first baseman}
        \item{playerId.2B}{The MLBAM of the second baseman}
        \item{playerId.3B}{The MLBAM of the third baseman}
        \item{playerId.SS}{The MLBAM of the shortstop}
        \item{playerId.LF}{The MLBAM of the left fielder}
        \item{playerId.CF}{The MLBAM of the center fielder}
        \item{playerId.RF}{The MLBAM of the right fielder}
        \item{batterPos}{The defensive postion of the current batter}
        \item{batterName}{The name of the current batter}
        \item{pitcherName}{The name of the current pitcher}
        \item{runsOnPlay}{The number of runs that were scored as a result of the plate appearance}
        \item{startOuts}{The number of outs at the beginning of the plate appearnace}
        \item{runsInInning}{The total number of runs scored in the half of an inning}
        \item{runsITD}{The total number of runs that have been scored in the half of an inning prior to the current plate appearance}
        \item{runsFuture}{The total number of runs that are scored in the half of an inning that were scored after the current plate appearance}
        \item{startCode}{Binary representation of the base runner configuration at the start of the plate appearance}
        \item{endCode}{Binary representation of the base runner configuration at the end of the plate appearance}
        \item{fielderId}{For a ball in play, the MLBAM id of the fielder who first touches the baseball. (NA for balls not in play) }
        \item{endCode}{Binary representation of the base runner configuration at the end of the plate appearance}
        \item{gameId}{The MLBAM id for this game}
        \item{isPA}{Boolean indicating if the plate appearance was a plate appearance}
        \item{isAB}{Boolean indicating if the plate appearance was an at bat}
        \item{isHit}{Boolean indicating if the plate appearance resulted in a hit}
        \item{isBIP}{Boolean indicating if the plate appearance resulted in a ball in play}
        \item{our.x}{}
        \item{out.y}{}
        \item{r}{}
        \item{theta}{}
   }}
\source{
\url{http://gd2.mlb.com/components/game/mlb/year_2012/month_05/day_14/}
}
\usage{
May14
}
\description{
A dataset containing play-by-play data from MLBAM GameDay and processed by
\code{openWAR} for all games on May 14, 2013.
}
\keyword{datasets}

