% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mosaic.do}
\alias{mosaic.do}
\title{Used internally in \code{shakeWAR}}
\usage{
mosaic.do(...)
}
\arguments{
\item{...}{arguments passed to \code{mosaic::do}.}
}
\description{
This is a hack to avoid warning about importing both mosaic::do and dplyr::do
}
