% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{openWARPlayers-class}
\alias{openWARPlayers-class}
\title{openWARPlayers}
\description{
A data.frame of players and their tabulated openWAR values.  The function \code{getWAR} returns an object of class \code{openWARPlayers}
}
\examples{
showClass("openWARPlayers")
}

