% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlays.R
\docType{class}
\name{openWARPlays-class}
\alias{openWARPlays-class}
\title{openWARPlays}
\description{
Contains the output from makeWAR() specific to openWAR
}
\examples{
showClass('openWARPlays')
}

