% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{panel.war}
\alias{panel.war}
\title{panel.war}
\usage{
panel.war(x, y, ...)
}
\arguments{
\item{x}{the x-coordinate}

\item{y}{the y-coordinate}

\item{...}{arguments passed from \code{\link{plot.openWARPlayers}}}
}
\description{
Display a season's worth of openWAR results
}
\details{
Given an \code{\link{openWARPlayers}} object, draw a plot displaying each player's RAA, WAR, and replacement
level shadow.
}
\examples{
\dontrun{
MayProcessed = makeWAR(May)
}
war = getWAR(MayProcessed$openWAR)
summary(war)
plot(war)
}

