% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{plot.do.openWARPlayers}
\alias{plot.do.openWARPlayers}
\title{Visualize resampled openWAR}
\usage{
\method{plot}{do.openWARPlayers}(x, playerIds = c(431151, 285079), ...)
}
\arguments{
\item{x}{A data.frame resulting from shakeWAR() of class \code{do.openWARPlayers}}

\item{playerIds}{A vector of valid MLBAM player IDs present in the data argument}

\item{...}{arguments passed to \code{\link{densityplot}}}
}
\value{
a faceted densityplot
}
\description{
Visualize resampled openWAR
}
\details{
Density Plots for openWAR estimates
}
\examples{


\dontrun{
openWAR.sim = shakeWAR(May)
plot(openWAR.sim, playerIds = c(431151, 502517, 408234, 285078, 518774, 285079))
}
}
