% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.GameDayPlays}
\alias{plot.GameDayPlays}
\title{plot.GameDayPlays}
\usage{
\method{plot}{GameDayPlays}(data, ...)
}
\arguments{
\item{data}{A GameDayPlays set with fields "our.x" and "our.y"}
}
\value{
an xyplot()
}
\description{
Visualize Balls in Play
}
\details{
Plots the balls in play from GameDay data. This function will plot (x,y)-coordinates
with a generic baseball field plotted in the background. Other lattice options can be passed
to xyplot().
}
\examples{
ds = getData()
plot(ds)
}

