% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{plot.openWARPlayers}
\alias{plot.openWARPlayers}
\title{Display a season's worth of openWAR results}
\usage{
\method{plot}{openWARPlayers}(x, ...)
}
\arguments{
\item{x}{A data.frame object of class \code{\link{openWARPlayers}}}

\item{...}{arguments passed to \code{\link{xyplot}}.}
}
\description{
Display a season's worth of openWAR results
}
\details{
Given an \code{\link{openWARPlayers}} object, draw a plot displaying each player's RAA, WAR, and replacement
level shadow.
}
\examples{

\dontrun{
MayProcessed = makeWAR(May)
}
war = getWAR(MayProcessed$openWAR)
summary(war)
plot(war)
}
