% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFielding.R
\name{plotFielding}
\alias{plotFielding}
\alias{plotFielding.glm}
\alias{plotFielding.lm}
\alias{plotFielding.formula}
\alias{plotFielding.default}
\alias{plotFielding.bkde2D}
\alias{panel.fielding}
\title{Plot fielding model for a player.}
\usage{
plotFielding(x, ...)

\method{plotFielding}{glm}(x, ...)

\method{plotFielding}{lm}(x, ...)

\method{plotFielding}{formula}(x, ...)

\method{plotFielding}{default}(x, ...)

\method{plotFielding}{bkde2D}(x, ...)

panel.fielding(x, y, z, ...)
}
\arguments{
\item{x}{a model object, currently \code{\link{lm}}, \code{\link{glm}}, and
\code{\link{bkde2D}} are
 supported. For the default method this is an x-coordinate to be passed to
 \code{\link{contourplot}}.}

\item{...}{arguments passed to \code{\link{panel.contourplot}}}

\item{y}{y-coordinate for default method passed to \code{\link{contourplot}}.}

\item{z}{z-coordinate for default method passed to \code{\link{contourplot}}.}
}
\value{
A contourplot object
}
\description{
Creates a plot of the fielding model for an individual player.
}
\details{
This is a convenience function for visually checking the fielding models. 
It provides a way to inspect the quality of the fielding models for individual
players, and was used to generate one of the figures used in the \code{\link{openWAR}}
paper.
}
\examples{
fmod <- getModelFieldingCollective(May)
plotFielding(fmod)

}
