\name{readData.gameday}
\alias{readData.gameday}
\title{readData.gameday}
\usage{
  readData.gameday(gd)
}
\arguments{
  \item{gd}{A gameday object}
}
\value{
  A gameday object
}
\description{
  Grabs XML files from MLBAM and converts to data.frames in
  memory
}
\details{
  Grabs XML files from MLBAM and converts to data.frames in
  memory
}
\examples{
gd = gameday()
getURLs.gameday(gd)
}

