% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shakeWAR.R
\name{shakeWAR}
\alias{shakeWAR}
\alias{shakeWAR.GameDayPlays}
\alias{shakeWAR.list}
\alias{shakeWAR.openWARPlays}
\title{shakeWAR}
\usage{
shakeWAR(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{list}(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{openWARPlays}(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{GameDayPlays}(data, N = 10, resample = "plays", ...)
}
\arguments{
\item{data}{An object that contains \code{\link{GameDayPlays}} data or data that has already been processed by \code{\link{makeWAR}}.}

\item{N}{the number of resamples (default 10)}

\item{resample}{An element of \code{c('plays', 'models', 'both')}. Currently only \code{plays}
and \code{both} are implemented. Only works with \code{shakeWAR.GameDayPlays} method.}

\item{...}{additional arguments passed to \code{shakeWAR} methods}
}
\value{
a data.frame with RAA values
}
\description{
resample a data.frame to obtain variance estimates for WAR


}
\details{
Resamples the rows of an MLBAM data set and calculates WAR values on the re-sampled data set.
This process is repeated many times to yield variance estimates of WAR.
}
\examples{
\dontrun{
res = shakeWAR(May, resample='plays', N=5)
summary(res)
}
\dontrun{
owar <- shakeWAR(MayProcessed, N = 5)
summary(owar)
}
\dontrun{
owar <- shakeWAR(MayProcessed$openWAR, N = 5)
summary(owar)
}
\dontrun{
owar <- shakeWAR(May, N = 7)
summary(owar)
owar <- shakeWAR(May, N = 2, resample = "both")
summary(owar)
}
}

