% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shakeWAR.R
\name{shakeWAR}
\alias{shakeWAR}
\alias{shakeWAR.GameDayPlays}
\alias{shakeWAR.list}
\alias{shakeWAR.openWARPlays}
\title{shakeWAR}
\usage{
shakeWAR(data, N = 10, ...)

\method{shakeWAR}{list}(data, N = 10, ...)

\method{shakeWAR}{openWARPlays}(data, N = 10, ...)

\method{shakeWAR}{GameDayPlays}(data, N = 10, ...)
}
\arguments{
\item{data}{An MLBAM data.frame}

\item{N}{the number of resamples (default 10)}

\item{...}{additional arguments passed to \code{shakeWAR} methods}

\item{resample}{An element of \code{c('plays', 'models', 'both')}. Currently only \code{plays}
and \code{both} are implemented.}
}
\value{
a data.frame with RAA values
}
\description{
resample a data.frame to obtain variance estimates for WAR
}
\details{
Resamples the rows of an MLBAM data set
}
\examples{
\dontrun{
res = shakeWAR(May, resample='plays', N=10)
summary(res)
}
owar <- shakeWAR(MayProcessed, N = 5)
summary(owar)
madeWAR <- makeWAR(May)
owar <- shakeWAR(madeWAR$openWAR, N = 5)
summary(owar)
owar <- shakeWAR(May, N = 7)
summary(owar)
\dontrun{
owar <- shakeWAR(May, N = 2, resample = "both")
summary(owar)
}
}

