% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shakeWAR.R
\name{shakeWAR}
\alias{shakeWAR}
\alias{shakeWAR.list}
\alias{shakeWAR.openWARPlays}
\alias{shakeWAR.GameDayPlays}
\title{Estimates the variance of openWAR}
\usage{
shakeWAR(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{list}(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{openWARPlays}(data, N = 10, resample = "plays", ...)

\method{shakeWAR}{GameDayPlays}(data, N = 10, resample = "plays", ...)
}
\arguments{
\item{data}{An object that contains \code{\link{GameDayPlays}} data or data that has already been processed by \code{\link{makeWAR}}.}

\item{N}{the number of resamples (default 10)}

\item{resample}{An element of \code{c('plays', 'models', 'both')}. Currently only \code{plays}
and \code{both} are implemented. Only works with \code{\link{shakeWAR.GameDayPlays}} method.}

\item{...}{additional arguments passed to \code{\link{shakeWAR}} methods}
}
\value{
a \code{do.\link{openWARPlayers}} data frame with RAA values
}
\description{
Resample to obtain variance estimates for WAR.
}
\details{
The general idea is to use resampling techniques to compute
variance estimates for openWAR. The \code{resample} argument controls 
how that resampling is done. 

For \code{resample = 'plays'}, \code{\link{makeWAR}} is run on the actual
plays passed in the \code{data} argument, if it hasn't been computed already.
The resulting RAA values (for each play) are an estimate of what actually happened. 
These RAA values are resampled \code{N} times. 

\code{resample = 'models'} is not currently implemented, but in future versions,
this option would allow you to resample the rows of the original data, run 
\code{\link{makeWAR}} on each resampled data set, collect all of the models
generated during that process, and then evaluate the original data using each
of the resampled models. This would allow you isolate the modeling error. 

For \code{resample = 'both'}, the plays are resampled, and \code{\link{makeWAR}}
is computed on each resampled data set. Note that this is both computationally
and memory intensive.
}
\examples{

\dontrun{
res <- shakeWAR(May, resample='plays', N=5)
summary(res)
}

\dontrun{
owar <- shakeWAR(MayProcessed, N = 5)
summary(owar)
}
\dontrun{
owar <- shakeWAR(MayProcessed$openWAR, N = 5)
summary(owar)
}
\dontrun{
owar <- shakeWAR(May, N = 7)
summary(owar)
owar <- shakeWAR(May, N = 2, resample = "both")
summary(owar)
}
}
