% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{summary.do.openWARPlayers}
\alias{summary.do.openWARPlayers}
\title{summary.do.openWARPlayers}
\usage{
\method{summary}{do.openWARPlayers}(object, n = 25, ...)
}
\arguments{
\item{object}{An object of class \code{\link{openWARPlayers}}}

\item{n}{the number of players to display}

\item{...}{currently ignored}
}
\description{
Summarize WAR
}
\details{
Summary of players' WAR
}
\examples{
\dontrun{
sim = shakeWAR(May)
summary(sim)
}
}

