% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{summary.do.openWARPlayers}
\alias{summary.do.openWARPlayers}
\title{Summarize resampled openWAR}
\usage{
\method{summary}{do.openWARPlayers}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{openWARPlayers}}}

\item{...}{currently ignored}
}
\description{
Summarize resampled openWAR
}
\details{
Summary of players' WAR
}
\examples{

\dontrun{
sim = shakeWAR(May)
summary(sim)
}
}
