% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{summary.GameDayPlays}
\alias{summary.GameDayPlays}
\title{summary.GameDayPlays}
\usage{
\method{summary}{GameDayPlays}(data)
}
\arguments{
\item{data}{A GameDayPlays data set}
}
\value{
nothing
}
\description{
Summarize MLBAM data
}
\details{
Prints information about the contents of an GameDayPlays data set.
}
\examples{
ds = getData()
summary(ds)
}

