\name{summary.GameDayPlays}
\alias{summary.GameDayPlays}
\title{summary.GameDayPlays}
\usage{
  summary.GameDayPlays(data)
}
\arguments{
  \item{data}{A GameDayPlays data set}
}
\value{
  nothing
}
\description{
  Summarize MLBAM data
}
\details{
  Prints information about the contents of an GameDayPlays
  data set.
}
\examples{
ds = getData()
summary(ds)
}

