% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{summary.GameDayPlays}
\alias{summary.GameDayPlays}
\title{Summarize a GameDayPlays data set}
\usage{
\method{summary}{GameDayPlays}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{GameDayPlays}}}

\item{...}{currently ignored}
}
\value{
A \code{\link{tbl_df}} of totals for each team
}
\description{
Summarize MLBAM data
}
\details{
Tabulates cumulative statistics by team for the contents of a \code{\link{GameDayPlays}} data set.
}
\examples{

summary(May)

}
