% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{summary.openWARPlayers}
\alias{summary.openWARPlayers}
\title{summary.openWARPlayers}
\usage{
\method{summary}{openWARPlayers}(object, n = 25, ...)
}
\arguments{
\item{object}{An object of class \code{'openWARPlayers'}}

\item{n}{the number of players to display}

\item{...}{currently ignored}
}
\description{
Summarize WAR among players
}
\details{
A summary of players' WAR
}
\examples{
war <- getWAR(MayProcessed$openWAR)
summary(war)
}

