% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWARPlayers.R
\name{summary.openWARPlayers}
\alias{summary.openWARPlayers}
\title{Summarize openWAR across players}
\usage{
\method{summary}{openWARPlayers}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{openWARPlayers}}}

\item{...}{currently ignored}
}
\description{
Summarize openWAR across players
}
\details{
A summary of players' WAR
}
\examples{

\dontrun{
MayProcessed = makeWAR(May)
}
war <- getWAR(MayProcessed$openWAR)
summary(war)
summary(openWAR2012)
summary(openWAR2013)
}
