% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GameDayPlays.R
\name{tabulate}
\alias{tabulate}
\alias{tabulate.GameDayPlays}
\title{tabulate}
\usage{
tabulate(data)

tabulate.GameDayPlays(data)
}
\arguments{
\item{data}{A GameDayPlays set}
}
\value{
A data.frame of seasonal totals for each team
}
\description{
Summarize MLBAM data
}
\details{
Tabulates Lahman-style statistics by team for the contents of a GameDayPlays data set.
}
\examples{
library(openWAR)
tabulate(May)
}

