% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tabulate}
\alias{tabulate}
\title{tabulate.GameDayPlays}
\usage{
tabulate(data)
}
\arguments{
\item{data}{A GameDayPlays set}
}
\value{
A data.frame of seasonal totals for each team
}
\description{
Summarize MLBAM data
}
\details{
Tabulates Lahman-style statistics by team for the contents of a GameDayPlays data set.
}
\examples{
ds = getData()
tabulate(ds)
}

