% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{updateGame}
\alias{updateGame}
\title{updateGame}
\usage{
updateGame(gameIds, data, ...)
}
\arguments{
\item{gameIds}{A character vector of valid MLBAM gameIds}

\item{data}{a data.frame returned by \code{\link{getData}}}

\item{...}{currently ignored}
}
\value{
a data.frame
}
\description{
Replaces data from a single game
}
\details{
Deletes, and then rbinds fresh information from a particular game.
}
\examples{
data(May)
dim(May)
tail(May)
newMay <- updateGame(gameIds = 'gid_2013_05_14_kcamlb_anamlb_1', data = May)
dim(newMay)
# the replaced games are now at the bottom of the data frame
tail(newMay)
}

