% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{updateGame}
\alias{updateGame}
\title{Update data for a single game from GameDay}
\usage{
updateGame(gameIds, data, ...)
}
\arguments{
\item{gameIds}{A character vector of valid MLBAM gameIds}

\item{data}{a data.frame returned by \code{\link{getData}}}

\item{...}{currently ignored}
}
\value{
A data.frame of class \code{\link{GameDayPlays}} consisting of play-by-play data
}
\description{
Replaces data from a single game.
}
\details{
Deletes and then appends fresh information from a particular game to the original data.
}
\examples{

data(May)
dim(May)
tail(May)
newMay <- updateGame(gameIds = 'gid_2013_05_14_kcamlb_anamlb_1', data = May)
dim(newMay)
# the replaced games are now at the bottom of the data frame
tail(newMay)
}
