% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameDay.R
\name{updateHalfInning}
\alias{updateHalfInning}
\title{Update action within a half-inning}
\usage{
updateHalfInning(dat)
}
\arguments{
\item{dat}{a \code{GameDayPlays} object containing information about a single inning.}
}
\value{
A gameday object
}
\description{
updateHalfInning
}
